[SqlAlchemy 1.4] Remove GenericModel
